requirements.txt
alembic==1.4.1
amqp==2.5.2
ansible==2.7.10
bcrypt==3.1.7
billiard==3.6.3.0
blinker==1.4
celery==4.4.1
cffi==1.14.0
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Avatars==0.2.2
Flask-CeleryExt==0.3.4
Flask-Cors==3.0.8
Flask-Limiter==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.8
limits==1.5.1
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
oss2==2.10.0
paramiko==2.7.1
Pillow==7.1.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3
redis==3.4.1
six==1.14.0
SQLAlchemy==1.3.13
vine==1.3.0
webargs==6.0.0
Werkzeug==1.0.0
zipp==3.0.0
